reply-generator-main/reply-generator-main/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
openai
python-dotenv
torch
numpy
